moved components from routes folder to components folder
